Divided the register component in pieces so it is easier to read and understand, modified code in the flux and routes to improve readibility :art:
